update speedPID and Stop Car

+++ user/include/control.h
@@ -20,7 +20,7 @@ extern PID_Calibration PID_adc;
 extern PID_Calibration PID_out_left;
 extern PID_Calibration PID_out_right;
 
-extern int TargetSpeed;	//目标速度，PID控制
+extern long TargetSpeed;	//目标速度，PID控制
 
 void Update_Motors(PID_State * left_state,PID_State * right_state,PID_State * gyro_state);
 void Set_Motors(long left, long right);
